refactor: transaction builder
